Extract calculation method from vm/lazy.ts
